Expanded UI system everything is derived from UIElem, and now has mouse event callbacks, as well a list of direct descendants. The descendants should be on top of the element visually
The collision function checks for a mouse pointer collision based on the geometry of the element, as well as all of its descendents, returning the first one to match

The draw function takes no arguments, so instead the color must be kept as state

The counter element is an example of a composite element, comprised of a frame with two buttons, where the buttons act directly on the internal value

TODO: mouse UI backend that checks for mouse collisions and sends the appropriate events like left click(only active on mouse up, maybe)  or wheel motion.

Added a solver control class that spawns a second thread so that the lengthy preparation and solution finding does no impede the user interface.

Added vec3.h and made it templated, with vec3 remaining the single precision float vector, while vec3i is the integer vector (and some of the operations don't make sense)